requirements.txt
# flask
# gunicorn
# langchain
# langchain_groq
# langchain_core
# numpy
# pandas
# python-dotenv



flask
gunicorn
langchain==0.1.16
langchain_groq
numpy
pandas
python-dotenv
